Unity Catalog changes from databricks-cli-private
* pylint fixes, version update
* Set version to '0.16.3.dev0.uc-int.1' (during previous master merge)
* Fix line-too-long errors from pylint.
* Split UC CLI commands (and API handlers, service class) into
  public and internal variants.
* Create mkdist.sh script for preparing distributions.
* Renamed 'get-recipient-share-perms' CLI as 'list-recipient-permissions'.
* Renamed 'list-tables-bulk' CLI as 'list-table-summaries'.
* Use 'list-share-permissions --name foo' and 'update-share-permissions --name foo'
  instead of 'get-permissions --share <foo>' and
  'update-permissions --share foo".
* Use 'hive_metastore' instead of 'main' as default-catalog-name in
  assign-metastore CLI.
* Support creating delta sharing catalog
* Fixed syntax, lint glitches
* Fix create-provider and update-provider namings
* Go back to using json_inline for UpdateExternalLocation (including
  the 'force' arg).
* Lint fixes.
* Unsplit the distribution (no more _internal.py src files or
  -int tarballs). Just create one UC distribution with all UC CLIs.
* For modularity, slit up the CLIs into metastore, catalog, etc.
  src files.
* Rename sharing -> delta_sharing
* delta-sharing improvements for SC-89547
* Use correct format
* Add 'unassign-metastore' command.
* Fix wording for 'assign-metastore' comments and help.
* mkdist script fixes
* Add new command to CLI
